FIX: Oppdaterer fagsakperiode ved innhenting av ny ungdomsprogramperiode. Bruker hele gjeldende ungdomsprogramperiode til utledning av nye vilkårsperioder som skal legges til/vurderes på nytt
