+++ domenetjenester/mottak/src/test/java/no/nav/ung/sak/mottak/dokumentmottak/FagsakperiodeUtlederTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import java.time.LocalDate;
 import java.util.List;
 
+import no.nav.ung.sak.ungdomsprogram.forbruktedager.FagsakperiodeUtleder;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
